reorganize elements, add field section helper
