+++ packages/blocks/src/index.spec.ts
@@ -1,6 +1,6 @@
 import * as index from './index';
 
-describe('The index exports', () => {
+describe('The index', () => {
   it('exports everything', () => {
     expect.assertions(1);
     expect(Object.keys(index)).toMatchSnapshot();
