+++ packages/blocks/src/index.ts
@@ -1,2 +1,3 @@
 export * from './blocks';
+export * from './elements';
 export * from './formatting';
